requirements.txt
alembic==1.0.5
autoenv==1.0.0
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
nltk==3.4
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
redis==3.0.1
requests==2.21.0
rq==0.13.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.7
SQLAlchemy==1.2.16
urllib3==1.24.1
Werkzeug==0.14.1
